Move `std::io::Error` and friends into `core` and `alloc`
