+++ library/core/src/io/mod.rs
@@ -8,22 +8,25 @@ mod util;
 
 #[unstable(feature = "core_io_borrowed_buf", issue = "117693")]
 pub use self::borrowed_buf::{BorrowedBuf, BorrowedCursor};
-#[unstable(feature = "core_io", issue = "154046")]
-pub use self::cursor::Cursor;
-#[unstable(feature = "core_io", issue = "154046")]
-pub use self::error::ErrorKind;
 #[unstable(feature = "raw_os_error_ty", issue = "107792")]
 pub use self::error::RawOsError;
-#[doc(hidden)]
-#[unstable(feature = "core_io_internals", reason = "exposed only for libstd", issue = "none")]
-pub use self::error::{
-    Custom, CustomOwner, OsFunctions, decode_error_kind, format_os_error, is_interrupted,
-    set_functions,
+#[unstable(feature = "io_const_error_internals", issue = "none")]
+pub use self::error::SimpleMessage;
+#[unstable(feature = "io_const_error", issue = "133448")]
+pub use self::error::const_error;
+#[unstable(feature = "core_io", issue = "154046")]
+pub use self::{
+    cursor::Cursor,
+    error::{Error, ErrorKind, Result},
+    io_slice::{IoSlice, IoSliceMut},
+    util::{Chain, Empty, Repeat, Sink, Take, empty, repeat, sink},
 };
-#[unstable(feature = "core_io", issue = "154046")]
-pub use self::io_slice::{IoSlice, IoSliceMut};
-#[unstable(feature = "core_io", issue = "154046")]
-pub use self::util::{Chain, Empty, Repeat, Sink, Take, empty, repeat, sink};
 #[doc(hidden)]
 #[unstable(feature = "core_io_internals", reason = "exposed only for libstd", issue = "none")]
-pub use self::util::{chain, take};
+pub use self::{
+    error::{
+        Custom, CustomOwner, OsFunctions, decode_error_kind, format_os_error, is_interrupted,
+        set_functions,
+    },
+    util::{chain, take},
+};
